+++ FireworkFlutterSDKSample/lib/routes.dart
@@ -21,6 +21,7 @@ import 'package:fw_flutter_sdk_example/screens/more/enable_custom_cta_click_call
 import 'package:fw_flutter_sdk_example/screens/more/enable_custom_share_url_callback_screen.dart';
 import 'package:fw_flutter_sdk_example/screens/more/enable_custom_share_url_screen.dart';
 import 'package:fw_flutter_sdk_example/screens/more/enable_link_interaction_click_callback_screen.dart';
+import 'package:fw_flutter_sdk_example/screens/more/enable_giveaway_terms_click_callback_screen.dart';
 import 'package:fw_flutter_sdk_example/screens/more/enable_on_video_playback_log.dart';
 import 'package:fw_flutter_sdk_example/screens/more/open_video_screen.dart';
 import 'package:fw_flutter_sdk_example/screens/more/set_share_base_url_screen.dart';
@@ -159,6 +160,11 @@ final routeMap = <String, FWRouteFactory>{
   '/enable_link_interaction_click_callback': (settings) {
     return MaterialPageRoute(
       builder: (context) => const EnableLinkInteractionClickCallbackScreen(),
+    );
+  },
+  '/enable_giveaway_terms_click_callback': (settings) {
+    return MaterialPageRoute(
+      builder: (context) => const EnableGiveawayTermsClickCallbackScreen(),
     );
   },
 };